@@ -32,7 +32,7 @@ namespace ISTC.FiveStage.Technology.JumpAspNetCore.AspNetCore_04
             });
 
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Latest);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
